Fix PR 471 ReviewGPT round 9 finding
