+++ packages/query/src/projection/wearable-summary-compose.ts
@@ -570,6 +570,7 @@ function wearableDatasetFromProjectedBundle(bundle: ProjectedWearableSummaryBund
 
   return {
     activitySessionAggregates,
+    metricSuppressionEvidence: [],
     metricCandidates,
     provenanceDiagnostics: [],
     rawMetricCandidates: metricCandidates,
