+++ packages/query/src/wearables/candidates.ts
@@ -20,6 +20,12 @@ import {
 } from "./origin.ts";
 import { formatProviderName } from "./provider-policy.ts";
 import {
+  normalizeResourceToken,
+  resolveSleepCandidateProvider,
+  sleepMetricAssociatedWithWindow,
+  sleepMetricMatchesWindow,
+} from "./sleep-association.ts";
+import {
   buildCandidateId,
   collectSortedDatesDesc,
   latestIsoTimestamp,
@@ -39,12 +45,25 @@ import type {
   WearableHeartRateZoneAggregate,
   WearableMetricCandidate,
   WearableMetricKey,
+  WearableMetricSuppressionEvidence,
   WearableMetricSelection,
   WearableProvenanceDiagnostic,
   WearableSleepWindowCandidate,
 } from "./types.ts";
 
 import { compareMetricCandidateByDateDesc, compareSleepWindowByDateDesc } from "./selection.ts";
+
+const APPLE_HEALTH_KIT_PROVIDER = "apple-health-kit";
+const JUNCTION_SLEEP_STAGE_SUMMARY_NORMALIZER_VERSION = "junction-sleep-stage-summary.v1";
+const JUNCTION_SLEEP_STAGE_CYCLE_FALLBACK_NORMALIZER_VERSION = "junction-sleep-stage-cycle-fallback.v1";
+const INVALID_ZERO_SLEEP_METRIC_SUPPRESSION_KEYS = new Map<WearableMetricKey, string>([
+  ["totalSleepMinutes", "total-sleep-minutes"],
+  ["sleepEfficiency", "sleep-efficiency"],
+  ["deepMinutes", "deep-sleep-minutes"],
+  ["lightMinutes", "sleep-light-minutes"],
+  ["remMinutes", "rem-sleep-minutes"],
+]);
+const INVALID_ZERO_SLEEP_METRICS = new Set<string>(INVALID_ZERO_SLEEP_METRIC_SUPPRESSION_KEYS.keys());
 
 export function collectWearableDataset(
   vault: VaultReadModel,
@@ -165,18 +184,143 @@ export function collectWearableDataset(
   const filteredSleepWindows = dedupedSleepWindows.filter((candidate) => matchesDateFilters(candidate.date, filters));
   const sleepStageAggregates = buildSleepStageAggregateCandidates(sleepStageCandidates, dedupedSleepWindows)
     .filter((candidate) => matchesDateFilters(candidate.date, filters));
+  const sanitizedMetricCandidates = sanitizeWearableMetricCandidates(rawMetricCandidates, dedupedSleepWindows);
   const metricCandidates = [
-    ...dedupeExactMetricCandidates(rawMetricCandidates).candidates,
+    ...dedupeExactMetricCandidates(sanitizedMetricCandidates.candidates).candidates,
     ...dedupeExactMetricCandidates(sleepStageAggregates).candidates,
   ].sort(compareMetricCandidateByDateDesc);
 
   return {
     activitySessionAggregates: buildActivitySessionAggregates(activitySessions),
+    metricSuppressionEvidence: sanitizedMetricCandidates.suppressionEvidence,
     metricCandidates,
     provenanceDiagnostics: [...provenanceDiagnostics.values()].sort(compareWearableProvenanceDiagnostics),
     rawMetricCandidates,
     sleepWindows: filteredSleepWindows,
   };
+}
+
+function sanitizeWearableMetricCandidates(
+  candidates: readonly WearableMetricCandidate[],
+  sleepWindows: readonly WearableSleepWindowCandidate[],
+): {
+  candidates: WearableMetricCandidate[];
+  suppressionEvidence: WearableMetricSuppressionEvidence[];
+} {
+  const invalidCandidates = collectInvalidAppleZeroSleepMetricCandidates(candidates, sleepWindows);
+  const invalidRecordIds = new Set(invalidCandidates.flatMap((candidate) => candidate.recordIds));
+
+  return {
+    candidates: candidates.filter((candidate) => !candidate.recordIds.some((recordId) => invalidRecordIds.has(recordId))),
+    suppressionEvidence: buildMetricSuppressionEvidence(invalidCandidates),
+  };
+}
+
+function collectInvalidAppleZeroSleepMetricCandidates(
+  candidates: readonly WearableMetricCandidate[],
+  sleepWindows: readonly WearableSleepWindowCandidate[],
+): WearableMetricCandidate[] {
+  const invalidCandidates: WearableMetricCandidate[] = [];
+
+  for (const window of sleepWindows) {
+    if (resolveSleepCandidateProvider(window) !== APPLE_HEALTH_KIT_PROVIDER) {
+      continue;
+    }
+
+    const windowCandidates = candidates.filter((candidate) => sleepMetricAssociatedWithWindow(candidate, window));
+    const awakeMinutes = firstPositiveMetricValue(windowCandidates, "awakeMinutes");
+    const invalidZeroTotalCandidates = windowCandidates.filter((candidate) =>
+      candidate.metric === "totalSleepMinutes" &&
+      candidate.value === 0 &&
+      isInvalidZeroSleepSummaryOwnedCandidate(candidate, window)
+    );
+    if (
+      invalidZeroTotalCandidates.length === 0 ||
+      awakeMinutes === null ||
+      window.durationMinutes <= awakeMinutes + 1
+    ) {
+      continue;
+    }
+
+    invalidCandidates.push(
+      ...windowCandidates.filter((candidate) =>
+        INVALID_ZERO_SLEEP_METRICS.has(candidate.metric) &&
+        candidate.value === 0 &&
+        (
+          isInvalidZeroSleepSummaryOwnedCandidate(candidate, window) ||
+          isLegacyInvalidZeroSleepSummaryCompanion(candidate, invalidZeroTotalCandidates)
+        )
+      ),
+    );
+  }
+
+  const invalidRecordIds = new Set<string>();
+  return invalidCandidates.filter((candidate) => {
+    const unseenRecordIds = candidate.recordIds.filter((recordId) => !invalidRecordIds.has(recordId));
+    for (const recordId of unseenRecordIds) {
+      invalidRecordIds.add(recordId);
+    }
+    return unseenRecordIds.length > 0;
+  });
+}
+
+function buildMetricSuppressionEvidence(
+  invalidCandidates: readonly WearableMetricCandidate[],
+): WearableMetricSuppressionEvidence[] {
+  const evidence = new Map<string, { date: string; metricKey: string; recordIds: string[] }>();
+
+  for (const candidate of invalidCandidates) {
+    const metricKey = INVALID_ZERO_SLEEP_METRIC_SUPPRESSION_KEYS.get(candidate.metric as WearableMetricKey);
+    if (!metricKey) {
+      continue;
+    }
+
+    const key = `${candidate.date}\0${metricKey}`;
+    const existing = evidence.get(key) ?? {
+      date: candidate.date,
+      metricKey,
+      recordIds: [],
+    };
+    existing.recordIds = uniqueStrings([...existing.recordIds, ...candidate.recordIds]);
+    evidence.set(key, existing);
+  }
+
+  return [...evidence.values()];
+}
+
+function isInvalidZeroSleepSummaryOwnedCandidate(
+  candidate: WearableMetricCandidate,
+  window: WearableSleepWindowCandidate,
+): boolean {
+  const normalizerVersion = normalizeResourceToken(candidate.dataOrigin?.normalizerVersion);
+  if (normalizerVersion === JUNCTION_SLEEP_STAGE_CYCLE_FALLBACK_NORMALIZER_VERSION) {
+    return false;
+  }
+
+  return sleepMetricMatchesWindow(candidate, window) ||
+    normalizerVersion === JUNCTION_SLEEP_STAGE_SUMMARY_NORMALIZER_VERSION;
+}
+
+function isLegacyInvalidZeroSleepSummaryCompanion(
+  candidate: WearableMetricCandidate,
+  invalidZeroTotalCandidates: readonly WearableMetricCandidate[],
+): boolean {
+  if (normalizeResourceToken(candidate.dataOrigin?.normalizerVersion)) {
+    return false;
+  }
+
+  return invalidZeroTotalCandidates.some((totalCandidate) =>
+    candidate.provider === totalCandidate.provider &&
+    wearableDataOriginKey(candidate.dataOrigin) === wearableDataOriginKey(totalCandidate.dataOrigin) &&
+    candidate.recordedAt === totalCandidate.recordedAt
+  );
+}
+
+function firstPositiveMetricValue(
+  candidates: readonly WearableMetricCandidate[],
+  metric: WearableMetricKey,
+): number | null {
+  return candidates.find((candidate) => candidate.metric === metric && candidate.value > 0)?.value ?? null;
 }
 
 export function buildActivitySessionAggregates(
